selftests: Add nvmf_loop.sh script to test NVMe over Fabrics (LOOP)
This patch adds a loopback selftest for the NVMf driver. To run this
test run the command (as root)

./nvmf_loop.sh

or

make run_tests

or

make check

in the tools/testing/selftest/nvmf folder. To show the help for the
underlying script run

./nvmf_loop.sh -h

This first test is very simple. It uses a loopback through the fabric
layer so it can be run on a single machine.

This selftest suite will be augmented over time.

For reference the results of this test on a VM running yielded the
following:

-----------------
running nvmf_loop
-----------------
testing target directly (reads)...
1048576+0 records in
1048576+0 records out
536870912 bytes (537 MB) copied, 0.473085 s, 1.1 GB/s
testing target directly (writes)...
1048576+0 records in
1048576+0 records out
536870912 bytes (537 MB) copied, 0.92303 s, 582 MB/s

testing via loopback (reads)...
1048576+0 records in
1048576+0 records out
536870912 bytes (537 MB) copied, 0.538461 s, 997 MB/s
testing via loopback (writes)...
1048576+0 records in
1048576+0 records out
536870912 bytes (537 MB) copied, 1.86372 s, 288 MB/s

Signed-off-by: Stephen Bates <stephen.bates@pmcs.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>