pci: Provide sensible irq vector alloc/free routines
Add a function to allocate and free a range of interrupt vectors, using
MSI-X, MSI or legacy vectors (in that order) based on the capabilities
of the underlying device and PCIe complex.

Additionally a new helper is provided to get the Linux IRQ number for
given device-relative vector so that the drivers don't need to allocate
their own arrays to keep track of the vectors for the multi vector
MSI-X case.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Alexander Gordeev <agordeev@redhat.com>