nvme-rdma: set queue count on reset / reconnect
Including updating the count in blk-mq.  Also replace the confusing
queue_count member with an online_queues one similar to PCIe that
tracks the number of queues that need to be torn down.

This also involves slightly reordering the reset code so that we
first initialize the newly created queues before re-initializing
the tagset.

Signed-off-by: Christoph Hellwig <hch@lst.de>