nvme-fabrics: add a generic NVMe over Fabrics library
Add a new nvme-fabrics module which contains transport independent NVMe
over Fabrics command.  For now this is code that sets up the Connect and
Propery/Get Fabrics commands and submits them using blk_execute_rq to the
lower layers, as well as a generic sysfs interface to add a new
controller.

Includes lots of fixes and improvements from:

    Ian Bakshan <ianb@mellanox.com>,
    Jay Freyensee <james.p.freyensee@intel.com>,
    Sagi Grimberg <sagig@mellanox.com>,
    Ming Lin <ming.l@ssi.samsung.com>

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Sagi Grimberg <sagig@mellanox.com>
Reviewed-by: Jay Freyensee <james.p.freyensee@intel.com>
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>