scsi_transport_sas: switch to bsg-lib for SMP passthrough
Simplify the SMP passthrough code by switching it to the generic bsg-lib
helpers that abstract away the details of the request code, and gets drivers
out of seeing struct scsi_request.

For the libsas host SMP code there is a small behavior difference in that
we now always clear the residual len for successfull commands, similar to the
three other SMP handler implementations.  Given that there is no partial
command handling in the host SMP handler this should not matter in practice.

WARNING: completely untested.

Signed-off-by: Christoph Hellwig <hch@lst.de>